+++ ui/components/SearchForTrainingsAndJobs/components/ChoiceColumn.js
@@ -1,7 +1,8 @@
-import React, { useState, useEffect } from "react";
+import React, { useEffect, useContext } from "react";
 import { useRouter } from "next/router";
-import { useScopeContext } from "../../../context/ScopeContext";
-import { useDispatch, useSelector } from "react-redux";
+import { ScopeContext } from "../../../context/ScopeContext";
+import { SearchResultContext } from "../../../context/SearchResultContextProvider";
+import { DisplayContext } from "../../../context/DisplayContextProvider";
 import distance from "@turf/distance";
 import { scrollToTop, scrollToElementInContainer, getItemElement } from "../../../utils/tools";
 import ItemDetail from "../../../components/ItemDetail/ItemDetail";
@@ -13,14 +14,6 @@ import pushHistory from "../../../utils/pushHistory";
 import dosearchImage from "public/images/dosearch.svg";
 import whispers from "../services/whispers.js";
 
-import {
-  setTrainings,
-  setSelectedItem,
-  setItemToScrollTo,
-  setFormValues,
-  setExtendedSearch,
-  setJobs,
-} from "../../../store/actions";
 import { flyToMarker, flyToLocation, closeMapPopups, setSelectedMarker } from "../../../utils/mapTools";
 
 const ChoiceColumn = ({
@@ -41,13 +34,20 @@ const ChoiceColumn = ({
   activeFilter,
   setActiveFilter,
 }) => {
-  const dispatch = useDispatch();
   const router = useRouter();
-  const scopeContext = useScopeContext();
-
-  const { trainings, jobs, selectedItem, itemToScrollTo, formValues /*, currentPage*/ } = useSelector(
-    (state) => state.trainings
-  );
+  const scopeContext = useContext(ScopeContext);
+  const {
+    trainings,
+    jobs,
+    setTrainings,
+    setJobs,
+    setSelectedItem,
+    selectedItem,
+    itemToScrollTo,
+    setItemToScrollTo,
+    setExtendedSearch,
+  } = useContext(SearchResultContext);
+  const { formValues, setFormValues } = useContext(DisplayContext);
 
   useEffect(() => {
     if (itemToScrollTo) {
@@ -55,7 +55,7 @@ const ChoiceColumn = ({
 
       if (itemElement) {
         scrollToElementInContainer("resultList", itemElement, 150, "auto");
-        dispatch(setItemToScrollTo(null));
+        setItemToScrollTo(null);
       }
     }
   });
@@ -71,7 +71,7 @@ const ChoiceColumn = ({
   const handleSelectItem = (item) => {
     flyToMarker(item, 12);
     closeMapPopups();
-    dispatch(setSelectedItem(item));
+    setSelectedItem(item);
 
     setSelectedMarker(item);
 
@@ -109,10 +109,10 @@ const ChoiceColumn = ({
   };
 
   const searchForJobsWithLooseRadius = async () => {
-    dispatch(setExtendedSearch(true));
+    setExtendedSearch(true);
     scrollToTop("choiceColumn");
 
-    dispatch(setJobs([]));
+    setJobs([]);
     const searchTimestamp = new Date().getTime();
     pushHistory({
       router,
@@ -126,13 +126,13 @@ const ChoiceColumn = ({
   };
 
   const searchOnNewCenter = async (newCenter, isTrainingSearch, isJobSearch) => {
-    dispatch(setExtendedSearch(false));
+    setExtendedSearch(false);
 
     scrollToTop("choiceColumn");
 
     formValues.location = newCenter;
 
-    dispatch(setFormValues(formValues));
+    setFormValues(formValues);
 
     // mise à jour des infos de distance des formations par rapport au nouveau centre de recherche
     if (isJobSearch) {
@@ -165,7 +165,7 @@ const ChoiceColumn = ({
       trainings[i].place.distance =
         Math.round(10 * distance(coordinates, [trainings[i].place.longitude, trainings[i].place.latitude])) / 10;
     }
-    dispatch(setTrainings(trainings));
+    setTrainings(trainings);
   };
 
   const getResultLists = () => {
@@ -236,17 +236,16 @@ const ChoiceColumn = ({
 
   return (
     <div id="choiceColumn" className={`choiceCol w-75 ${shouldShowWelcomeMessage ? "c-choicecolumn__nosearch" : ""}`}>
-
-        {isLoading ? (
-          <LoadingScreen />
-        ) : (
-          <>
-            {getInitialDesktopText()}
-            {getSearchForm()}
-            {trainings.length === 0 && isJobSearchLoading ? <div></div> : getResultLists()}
-            {getSelectedItemDetail()}
-          </>
-        )}
+      {isLoading ? (
+        <LoadingScreen />
+      ) : (
+        <>
+          {getInitialDesktopText()}
+          {getSearchForm()}
+          {trainings.length === 0 && isJobSearchLoading ? <div></div> : getResultLists()}
+          {getSelectedItemDetail()}
+        </>
+      )}
     </div>
   );
 };