+++ ui/context/ScopeContext.js
@@ -1,6 +1,6 @@
-import React from "react";
+import React, { createContext } from "react";
 
-const ScopeContext = React.createContext({
+export const ScopeContext = createContext({
   isTraining: true,
   isJob: true,
 });
@@ -8,6 +8,3 @@ const ScopeContext = React.createContext({
 export function ScopeContextProvider(props) {
   return <ScopeContext.Provider value={props.value}>{props.children}</ScopeContext.Provider>;
 }
-
-export const useScopeContext = () => React.useContext(ScopeContext);
-export default ScopeContext;